requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
aiohttp>=3.9.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=4.9.0
tqdm>=4.66.0
py7zr>=0.20.0
rarfile>=4.1
trimesh>=4.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
